requirements.txt
streamlit>=1.35.0
openpyxl>=3.1.2
Pillow>=10.0.0
python-dotenv>=1.0.0
